requirements.txt
Pillow
transformers
pyyaml
